requirements.txt
requests==2.31.0
openai>=1.0.0
tweepy==4.14.0
python-dotenv==1.0.0
requests-oauthlib==1.3.1
PySocks==1.7.1
urllib3==2.0.7 